conditional caption area classes depending on view type
in an attempt to move the blacklight classes out of the scss into the
relevant code, i moved the "bg-dark" and "text-white" into the
document_component. upon doing this, i realized that this affected all
of the views, not just masonry. the css targets specific views, but each
view uses the document component. in order to use these blacklight
classes JUST for masonry, we need to conditionally determine the classes
to render.